+++ src/main/java/com/yourmode/yourmodebackend/domain/survey/dto/SurveyOptionResponseDto.java
@@ -15,5 +15,5 @@ public class SurveyOptionResponseDto {
     @Schema(description = "옵션 내용", example = "스트레이트형")
     private String content;
     @Schema(description = "체형 타입 ID", example = "1")
-    private Long bodyTypeId;
+    private Integer bodyTypeId;
 }
