+++ src/main/java/com/yourmode/yourmodebackend/domain/user/repository/BodyTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BodyTypeRepository extends JpaRepository<BodyType, Long> {
+public interface BodyTypeRepository extends JpaRepository<BodyType, Integer> {
     Optional<BodyType> findByName(String name);
 }
